+++ react/components/AccordionFilterContainer.js
@@ -31,7 +31,7 @@ const AccordionFilterContainer = ({
   priceRange,
   loading,
   onClearFilter,
-  showClearByFilterOnMobile,
+  showClearByFilter,
   updateOnFilterSelectionOnMobile,
 }) => {
   const [openItem, setOpenItem] = useState(null)
@@ -157,7 +157,7 @@ const AccordionFilterContainer = ({
                 onOpen={handleOpen(title)}
                 onFilterCheck={onFilterCheck}
                 onClearFilter={onClearFilter}
-                showClearByFilterOnMobile={showClearByFilterOnMobile}
+                showClearByFilter={showClearByFilter}
               />
             )
         }
@@ -194,7 +194,7 @@ AccordionFilterContainer.propTypes = {
   /** Clear filter function */
   onClearFilter: PropTypes.func,
   /** Whether a clear button that clear all options in a specific filter should appear beside the filter's name (true) or not (false). */
-  showClearByFilterOnMobile: PropTypes.bool,
+  showClearByFilter: PropTypes.bool,
   /** Wether the search will be updated on facet selection (`true`) or not (`false`) when the user is on mobile. */
   updateOnFilterSelectionOnMobile: PropTypes.bool,
 }